unit-rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.27
# chromadb==0.3.29  # Disabled - conflicts with fastapi
openai==1.3.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0
pydantic>=1.10,<2.0  # Compatible with all packages
pydantic-settings
sentence-transformers==2.2.2
numpy
pandas
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
